replay viewer: add 1/2x playback speed; Space pauses in the league shell (#1)
The replay playhead gains a 0.5x speed: command '5' selects the
ReplayHalfSpeed sentinel (speed == 0) on the ReplaySession, advance()
spends one tick every other frame via a halfPhase parity bit toggled as
its first statement, '-' now floors at 1/2x (1 div 2 == 0) and '+'
climbs back to 1x. The chrome's "sp" field carries the float
displaySpeed() -- 0.5 while at half (FrameContext.speed is a float now;
the live server sends 1.0) -- and wire_constants prepends 0.5 to the
emitted speeds array, so chrome_common builds a 0.5x chip mapped to
command '5' (its raw file:// SPEEDS fallback carries 0.5 too). The
lull-boost path keeps its full catch-up pace at half speed. The chrome
keeps reading window.CTF_WIRE: this repo deliberately emits the
CTF_WIRE=GEN_WIRE alias and the Dockerfile pins both lines.

The league shell (client/league_replayer.html -> dist/league.html)
forwards Space down the gen-shell command channel so play/pause works
while the shell chrome holds focus -- keydown never crosses the iframe
boundary. The board page (client/replay_broadcast.html -> dist/
index.html) already binds Space and relays the '1'..'9' digits, so the
new '5' works there with zero page edits.

tests: test_gen_replay grows a half-speed suite (sentinel selection,
0.5 display, tick-every-other-frame parity, the '-'/'+' floor and
exit); test_gen_viewer re-pins the patched chrome bytes (sha256 +
length) and pins the 0.5 chip map and SPEEDS fallback.

+++ src/generals/broadcast.nim
@@ -6,7 +6,8 @@
 ## `lead` series keep their structure; the fields are this game's.
 ##
 ## The INHERITED keys are unchanged, because `client/chrome_common.js` is
-## byte-for-byte the starter's and reads them: `t`, `mt`, `ph`, `lob`, `sp`,
+## the starter's (plus the fleet-wide 0.5x patch) and reads them: `t`, `mt`,
+## `ph`, `lob`, `sp`,
 ## `mx`, `st`, `lp`, `sk`, `ff`, `en`, `mm`, `teams`, `roster`, `events`,
 ## `lead`, `lulls`, `beats`, `over`, `pov`, `pl`.
 
@@ -19,7 +20,9 @@ type
     startTick*: int
     maxTick*: int
     playing*: bool
-    speed*: int
+    speed*: float
+      ## What the chrome shows: 0.5 while at the replay-only half speed,
+      ## else the integer multiplier.
     loop*: bool
     skipLulls*: bool
     fastForward*: bool
